% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airplane_emissions.R
\name{airplane_emissions}
\alias{airplane_emissions}
\title{Calculate Airplane Emissions}
\usage{
airplane_emissions(
  departing_airport_code,
  arriving_airport_code,
  via = NULL,
  num_people = 1,
  radiative_force = TRUE,
  round_trip = FALSE,
  class = "economy"
)
}
\arguments{
\item{departing_airport_code}{Takes a three-letter IATA code corresponding to an airport. Can check}

\item{arriving_airport_code}{Takes a three-letter IATA code corresponding to an airport.}

\item{via}{Optional. Takes a vector containing three-letter IATA codes corresponding to airports.}

\item{num_people}{Number of people taking the flight. Takes a single numerical value.}

\item{radiative_force}{Whether radiative force should be taken into account. Emissions from airplanes at higher altitudes impact climate change more than at ground level, radiative forcing accounts for this. Recommended.}

\item{round_trip}{Whether the flight is one-way or return.}

\item{class}{Class flown in. Options are "economy", "premium economy", "business", and "first".}
}
\value{
Returns CO2 emissions in tonnes.
}
\description{
A function that calculates CO2 emissions between airports. Distances are calculated using the airport_distance function in the airportr package.
}
\examples{
airplane_emissions("YVR","YYZ")
airplane_emissions("LHR", "NBO", "AMS")
}
