% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport_finder.R
\name{airport_finder}
\alias{airport_finder}
\title{Find the airport code for an airport}
\usage{
airport_finder(name, city, IATA_code, ICAO_code)
}
\arguments{
\item{name}{Name of the airport}

\item{city}{City the airport is in}

\item{IATA_code}{The IATA code}

\item{ICAO_code}{The ICAO code}
}
\value{

}
\description{
Find the airport code for an airport
}
