% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caRbon.R
\docType{package}
\name{caRbon}
\alias{caRbon}
\title{\code{caRbon-package} package}
\description{
Here
}
\seealso{
here
See the README on \href{https://github.com/lilyclements/caRbon/blob/update/README.md}{GitHub}
}
\author{
here
}
