% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caRbon.R, R/carbonr.R
\docType{package}
\name{carbonr}
\alias{carbonr}
\title{\code{carbonr-package} package}
\description{
Here

Here
}
\seealso{
here
See the README on \href{https://github.com/lilyclements/carbonr/blob/update/README.md}{GitHub}

here
See the README on \href{https://github.com/lilyclements/carbonr/blob/update/README.md}{GitHub}
}
\author{
here

here
}
