% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_emissions.R
\name{clinical_theatre_emissions}
\alias{clinical_theatre_emissions}
\title{Clinical Emissions}
\usage{
clinical_theatre_emissions(
  wet_clinical_waste,
  wet_clinical_waste_unit = c("tonnes", "kg"),
  desflurane = 0,
  sevoflurane = 0,
  isoflurane = 0,
  methoxyflurane = 0,
  N2O = 0,
  propofol = 0,
  water_supply = 0,
  water_trt = TRUE,
  water_unit = c("cubic metres", "million litres"),
  electricity_kWh = 0,
  electricity_TD = TRUE,
  electricity_WTT = TRUE,
  heat_kWh = 0,
  heat_TD = TRUE,
  heat_WTT = TRUE,
  glass = 0,
  board = 0,
  mixed = 0,
  paper = 0,
  average = 0,
  average_film = 0,
  average_rigid = 0,
  HDPE = 0,
  LDPE = 0,
  LLDPE = 0,
  PET = 0,
  PP = 0,
  PS = 0,
  PVC = 0,
  glass_WD = 0,
  glass_waste_disposal = c("Closed-loop", "Combustion", "Landfill", "Open-loop"),
  board_WD = 0,
  mixed_WD = 0,
  paper_WD = 0,
  fridges = 0,
  freezers = 0,
  electric_waste_disposal = c("Landfill", "Open-loop"),
  electrical_units = c("kg", "tonnes"),
  paper_waste_disposal = c("Closed-loop", "Combustion", "Composting", "Landfill"),
  average_WD = 0,
  average_film_WD = 0,
  average_rigid_WD = 0,
  HDPE_WD = 0,
  LDPE_WD = 0,
  LLDPE_WD = 0,
  PET_WD = 0,
  PP_WD = 0,
  PS_WD = 0,
  PVC_WD = 0,
  glass_units = c("kg", "tonnes"),
  paper_units = c("kg", "tonnes"),
  plastic_units = c("kg", "tonnes")
)
}
\arguments{
\item{wet_clinical_waste}{Amount of (wet) clinical waste that is usually incinerated.}

\item{wet_clinical_waste_unit}{Unit for \code{wet_clinical_waste} variable. Options are \code{"tonnes"} or \code{"kg"}.}

\item{desflurane}{Amount of desflurane used in KG (default: 0).}

\item{sevoflurane}{Amount of sevoflurane used in KG (default: 0).}

\item{isoflurane}{Amount of isoflurane used in KG (default: 0).}

\item{methoxyflurane}{Amount of methoxyflurane used in KG (default: 0).}

\item{N2O}{Amount of nitrous oxide (N2O) used in KG (default: 0).}

\item{propofol}{Amount of propofol used in KG (default: 0).}

\item{water_supply}{numeric, amount of water in the given unit.}

\item{water_trt}{logical, include treatment emissions (default TRUE).}

\item{water_unit}{"cubic metres" or "million litres".}

\item{electricity_kWh}{numeric kWh consumed.}

\item{electricity_TD}{logical, include T&D losses (default TRUE).}

\item{electricity_WTT}{logical, include WTT for electricity (default TRUE).}

\item{heat_kWh}{numeric kWh of heat/steam (onsite; excludes district).}

\item{heat_TD}{logical, include district heat distribution losses (default TRUE).}

\item{heat_WTT}{logical, include WTT for heat/steam (default TRUE).}

\item{glass}{Numeric tonnage of glass (material use).}
}
\value{
Returns CO2e emissions in tonnes.
}
\description{
Calculate carbon-equivalent emissions following an operating theatre
}
\examples{
\donttest{
 # clinical theatre emissions for the default options, with 100kg of wet clinical waste
 clinical_theatre_emissions(wet_clinical_waste = 100, wet_clinical_waste_unit = "kg")
}
}
