% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree_conversion.R
\name{degree_conversion}
\alias{degree_conversion}
\title{Convert degrees to radians}
\usage{
degree_conversion(deg)
}
\arguments{
\item{deg}{Degree value to convert.}
}
\value{
Value in radians.
}
\description{
Convert degrees to radians.
}
\examples{
# Convert 90 degrees into radians
degree_conversion(90)
}
