% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance_calc.R
\name{distance_calc}
\alias{distance_calc}
\title{Distance calculator}
\usage{
distance_calc(lat1, lat2, long1, long2)
}
\arguments{
\item{lat1}{Latitude of the first location.}

\item{lat2}{Latitude of the second location.}

\item{long1}{Longitude of the first location.}

\item{long2}{Longitude of the second location.}
}
\value{
Distance between locations in miles
}
\description{
Calculate distances between locations in miles using the Haversine formula. This is calculated as the crow flies.
}
\examples{
# Distance between the London Eye and the Eiffel Tower in miles
distance_calc(51.5033, 48.8584, 0.1196, 2.2945)
}
