% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotel_emissions.R
\name{hotel_emissions}
\alias{hotel_emissions}
\title{Calculate Emissions from a hotel stay}
\usage{
hotel_emissions(location = "UK", nights = 1, rooms = 1)
}
\arguments{
\item{location}{Location of the hotel stay. Current accepted locations are UK, Argentina, Australia, Austria, Belgium, Brazil, Canada, Chile, China, Columbia, Costa Rica, Czechia, Egypt, Fiji, France, Germany, Greece, Hong Kong, India, Indonesia, Ireland, Israel, Italy, Japan, Jordan, Korea, Macau, Malaysia, Maldives, Mexico, Netherlands, New Zealand, Panama, Peru, Philippines, Poland, Portugal, Qatar, Romania, Russia, Saudi Arabia, Singapore, Slovakia, South Africa, Spain, Switzerland, Taiwan, Thailand, Turkey, United Arab Emirates, United States, Vietnam, Average.}

\item{nights}{Number of nights stayed in the hotel.}

\item{rooms}{Number of rooms used in the hotel.}
}
\value{
Tonnes of CO2 emissions for a stay in a hotel.
}
\description{
Indirect emissions from a stay at a hotel. Values to calculate emissions are from UK government 2021 report.
}
\examples{
# Emissions for a two night stay in Fiji.
hotel_emissions(location = "Fiji", nights = 2)
}
