% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rail_emissions.R
\name{rail_emissions}
\alias{rail_emissions}
\title{Calculate Train Journey Emissions}
\usage{
rail_emissions(
  from,
  to,
  via = NULL,
  num_people = 1,
  times_journey = 1,
  round_trip = FALSE
)
}
\arguments{
\item{from}{Station departing from.}

\item{to}{Station arriving to}

\item{via}{Optional. Takes a vector containing the stations the train travels via.}

\item{num_people}{Number of people taking the journey. Takes a single numerical value.}

\item{times_journey}{Number of times the journey is taken.}

\item{round_trip}{Whether the journey is one-way or return.}
}
\value{
Returns CO2 emissions in tonnes for the train journey.
}
\description{
A function that calculates CO2 emissions between train stations in the UK.
}
\examples{
# Emissions for a train journey between Southampton Central and Manchester Piccadilly Station
rail_emissions("Southampton Central", "Manchester Piccadilly")
# Emissions for a train journey between Bristol Temple Meads and London Paddington
# via Bath, Swindon, and Reading
rail_emissions("Bristol Temple Meads", "London Paddington", via = c("Bath Spa", "Swindon", "Reading"))
}
