% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaport_lookup.R
\name{seaport_lookup}
\alias{seaport_lookup}
\title{Check the Code or Name of a Seaport}
\usage{
seaport_lookup(
  country_name = NULL,
  city = NULL,
  distance = 0.1,
  ignore.case = FALSE
)
}
\arguments{
\item{country_name}{Name of the country.}

\item{city}{Name of the city.}

\item{distance}{Maximum distance allowed for a match between the country/city given, and that of the value in the data set.}

\item{ignore.case}{If \code{FALSE}, the check for city/country is case-sensitive. If \code{TRUE}, case is ignored.}
}
\value{
Data frame containing the latitude, longitude, country name, country code, port code, and city name.
}
\description{
Check the Code or Name of a Seaport
}
\examples{
# Look up the city of Aberdeen to find the port_code for it
seaport_lookup(city = "Aberdeen")
# Search for a country and city and it finds matches
seaport_lookup(country_name = "United", city = "borunemouth", ignore.case = TRUE)
}
