% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaports.R
\docType{data}
\name{seaports}
\alias{seaports}
\title{Dataset of different seaports}
\format{
A data frame with 8736 rows and 7 variables:
\describe{
\item{country}{Name of the country with the port}
\item{city}{Name of the city with the port}
\item{country_code}{Code of the country name}
\item{port_code}{Code of the city port}
\item{latitude}{Latitude of the port}
\item{longitude}{Longitude of the port}
}
}
\source{
\url{https://www.kaggle.com/therohk/world-seaport-airport-dataset-and-codes/script}
}
\usage{
data(seaports)
}
\description{
A dataset containing the country, country code, city, city code, and coordinates of seaports
}
\keyword{datasets}
