% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_emissions.R
\name{secondary_emissions}
\alias{secondary_emissions}
\title{Secondary
source: carbonfootprint.com for these. They use DEFRA 2017 Supply Chain Factors
Emissions caused through the manufacture, delivery, and disposal of products and services we buy}
\usage{
secondary_emissions(item, cost, number_bought = 1, diet = "medium meat")
}
\description{
Secondary
source: carbonfootprint.com for these. They use DEFRA 2017 Supply Chain Factors
Emissions caused through the manufacture, delivery, and disposal of products and services we buy
}
