% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_CPI.R
\name{valid_jurisdictions}
\alias{valid_jurisdictions}
\title{Check which jurisdictions are in the Carbon Credits data}
\usage{
valid_jurisdictions(jurisdiction = NULL, period = FALSE)
}
\arguments{
\item{jurisdiction}{(optional) A character string specifying the jurisdiction to filter the data by.}

\item{period}{(logical) If TRUE, include the Period column in the output data frame.}
}
\value{
A vector or data frame containing the information.
}
\description{
Find jurisdictions available in the Carbon Credits data. If a jurisdiction is specified, find the years associated with that jurisdiction.
}
\examples{
which_jur <- valid_jurisdictions()

which_years <- valid_jurisdictions(jurisdiction = "Switzerland")
which_years_and_period <- valid_jurisdictions(jurisdiction = "Switzerland", period = TRUE)
}
