% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airplane_emissions.R
\name{airplane_emissions}
\alias{airplane_emissions}
\title{Calculate CO2e emissions from a airplane journey}
\usage{
airplane_emissions(
  from,
  to,
  via = NULL,
  num_people = 1,
  radiative_force = TRUE,
  round_trip = FALSE,
  class = c("economy", "premium economy", "business", "first")
)
}
\arguments{
\item{from}{Takes a three-letter IATA code corresponding to an airport. Can check the IATA code by the \code{airport_finder} function.}

\item{to}{Takes a three-letter IATA code corresponding to an airport. Can check the IATA code by the \code{airport_finder} function.}

\item{via}{Optional. Takes a vector containing three-letter IATA codes corresponding to airports.}

\item{num_people}{Number of people taking the flight. Takes a single numerical value.}

\item{radiative_force}{logical. Whether radiative force should be taken into account. Recommended \code{TRUE}. Emissions from airplanes at higher altitudes impact climate change more than at ground level, radiative forcing accounts for this.}

\item{round_trip}{logical. Whether the flight is one-way or return.}

\item{class}{Class flown in. Options are \code{c("economy", "premium economy", "business", "first")}.}
}
\value{
Returns CO2e emissions in tonnes.
}
\description{
A function that calculates CO2e emissions between airports. Distances are calculated using the airport_distance function in the airportr package.
}
\examples{
# Emissions for a flight between Vancouver and Toronto
airplane_emissions("YVR","YYZ")
# Emissions for a flight between London Heathrow and Kisumu Airport, via Amsterdam and Nairobi
airplane_emissions("LHR", "KIS", via = c("AMS", "NBO"))
}
