% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbonr.R
\docType{package}
\name{carbonr}
\alias{carbonr}
\title{\code{carbonr-package} package}
\description{
Here
}
\seealso{
here
See the README on \href{https://github.com/IDEMSInternational/carbonr/blob/main/README.md}{GitHub}
}
\author{
here
}
