% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbonr.R
\docType{package}
\name{carbonr}
\alias{carbonr-package}
\alias{carbonr}
\title{carbonr: Calculate Carbon-Equivalent Emissions}
\description{
The \code{carbonr} package provides a flexible tool for calculating carbon-equivalent emissions.
}
\seealso{
See the README on \href{https://github.com/IDEMSInternational/carbonr/blob/main/README.md}{GitHub}
}
\author{
\strong{Maintainer}: Lily Clements \email{lily@idems.international} (\href{https://orcid.org/0000-0001-8864-0552}{ORCID})

}
