% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/office_emissions.R
\name{office_emissions}
\alias{office_emissions}
\title{CO2e emissions from office work.}
\usage{
office_emissions(
  specify = FALSE,
  num_people = 1,
  num_wfh,
  electricity_kwh,
  kgco2e = 0.21233,
  butane,
  butane_units,
  CNG,
  CNG_units,
  LPG,
  LPG_units,
  LNG,
  LNG_units,
  natural_gas,
  natural_gas_units = "kwh",
  gas,
  gas_units,
  propane,
  propane_units,
  burning_oil,
  burning_oil_units,
  diesel,
  diesel_units,
  coal,
  coal_units,
  wood,
  wood_units
)
}
\arguments{
\item{specify}{Whether an average should be used, or if they wish to specify the amount of energy used.}

\item{num_people}{Number of people in the office.}

\item{num_wfh}{Number of people working from home.}

\item{electricity_kwh}{}

\item{kgco2e}{}

\item{butane}{}

\item{butane_units}{units that the gas is given in. Options are \code{"tonnes"}, \code{"litres"}, \code{"kwh"}.}

\item{CNG}{}

\item{CNG_units}{todo: can I give this with other _units ones?}

\item{LPG}{}

\item{LPG_units}{}

\item{LNG}{}

\item{LNG_units}{}

\item{natural_gas}{}

\item{natural_gas_units}{}

\item{gas}{}

\item{gas_units}{}

\item{propane}{}

\item{propane_units}{}

\item{burning_oil}{}

\item{burning_oil_units}{}

\item{diesel}{}

\item{diesel_units}{}

\item{coal}{}

\item{coal_units}{}

\item{wood}{}

\item{wood_units}{}
}
\value{

}
\description{
CO2e emissions from office work.
}
