% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/office_emissions.R
\name{office_emissions}
\alias{office_emissions}
\title{CO2e emissions from office work.}
\usage{
office_emissions(
  specify = FALSE,
  num_people = 1,
  num_wfh,
  electricity_kwh,
  kgco2e = 0.21233,
  butane,
  butane_units,
  CNG,
  CNG_units,
  LPG,
  LPG_units,
  LNG,
  LNG_units,
  natural_gas,
  natural_gas_units = "kwh",
  gas,
  gas_units,
  propane,
  propane_units,
  burning_oil,
  burning_oil_units,
  diesel,
  diesel_units,
  coal,
  coal_units,
  wood,
  wood_units
)
}
\arguments{
\item{specify}{Whether an average should be used, or if they wish to specify the amount of energy used.}

\item{num_people}{Number of people in the office.}

\item{num_wfh}{Number of people working from home.}

\item{electricity_kwh}{TODO}

\item{kgco2e}{TODO}

\item{butane}{TODO}

\item{butane_units}{units that the gas is given in. Options are \code{"tonnes"}, \code{"litres"}, \code{"kwh"}.}

\item{CNG}{TODO}

\item{CNG_units}{todo: can I give this with other _units ones?}

\item{LPG}{TODO}

\item{LPG_units}{TODO}

\item{LNG}{TODO}

\item{LNG_units}{TODO}

\item{natural_gas}{TODO}

\item{natural_gas_units}{TODO}

\item{gas}{TODO}

\item{gas_units}{TODO}

\item{propane}{TODO}

\item{propane_units}{TODO}

\item{burning_oil}{TODO}

\item{burning_oil_units}{TODO}

\item{diesel}{TODO}

\item{diesel_units}{TODO}

\item{coal}{TODO}

\item{coal_units}{TODO}

\item{wood}{TODO}

\item{wood_units}{TODO}
}
\value{
TODO
}
\description{
CO2e emissions from office work.
}
\examples{
TODO
}
