% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/office_emissions.R
\name{office_emissions}
\alias{office_emissions}
\title{Office emissions}
\usage{
office_emissions(
  specify = FALSE,
  num_people = 1,
  num_wfh = 0,
  electricity_kwh = 0,
  heat_kwh = 0,
  water_m3 = 0,
  include_TD = TRUE,
  include_WTT = TRUE,
  water_trt = TRUE,
  natural_gas = 0,
  burning_oil = 0,
  coal_domestic = 0,
  wood_log = 0,
  wood_chips = 0,
  natural_gas_units = c("kwh", "cubic metres", "tonnes"),
  burning_oil_units = c("kwh", "litres", "tonnes"),
  coal_domestic_units = c("kwh", "tonnes"),
  wood_log_units = c("kwh", "tonnes"),
  wood_chips_units = c("kwh", "tonnes")
)
}
\arguments{
\item{specify}{logical. Whether an average should be used, or if the amount of energy used will be specified.}

\item{num_people}{Number of people in the office.}

\item{num_wfh}{Number of people working from home.}

\item{electricity_kwh}{Electricity used in kWh.}

\item{heat_kwh}{Amount of heat/steam used for heating purposes in kWh.}

\item{water_m3}{Amount of water used in cubic-metres.}

\item{include_TD}{logical. Recommended \code{TRUE}. Whether to include emissions associated with grid losses.}

\item{include_WTT}{logical. Recommended \code{TRUE}. Whether to include emissions associated with extracting, refining, and transporting fuels.}

\item{water_trt}{logical. Recommended \code{TRUE}. Whether to include emissions associated with water treatment for used water.}

\item{natural_gas}{amount used. Standard natural gas received through the gas mains grid network in the UK.}

\item{burning_oil}{amount used. Main purpose is for heating/lighting on a domestic scale (also known as kerosene).}

\item{coal_domestic}{amount used. Coal used domestically.}

\item{wood_log}{amount used.}

\item{wood_chips}{amount used.}

\item{natural_gas_units}{units that the gas is given in. Options are \code{"tonnes"}, \code{"cubic metres"}, \code{"kwh"}.}

\item{burning_oil_units}{units that the fuel is given in. Options are \code{"tonnes"}, \code{"litres"}, \code{"kwh"}.}

\item{coal_domestic_units}{units that the fuel is given in. Options are \code{"kwh"}, \code{"tonnes"}.}

\item{wood_log_units}{units that the biomass is given in. Options are \code{"kwh"}, \code{"tonnes"}.}

\item{wood_chips_units}{units that the biomass is given in. Options are \code{"kwh"}, \code{"tonnes"}.}
}
\value{
Returns CO2e emissions in tonnes.
}
\description{
Further options are given in the \code{raw_fuels} function.
}
\examples{
# TODO
office_emissions(specify = TRUE, electricity_kwh = 200,
                 heat_kwh = 100, water_m3 = 100, include_TD = FALSE)

}
\references{
Descriptions from 2021 UK Government Report: https://www.gov.uk/government/publications/greenhouse-gas-reporting-conversion-factors-2021
}
