% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaport_lookup.R
\name{seaport_lookup}
\alias{seaport_lookup}
\title{Check the code or name of a seaport}
\usage{
seaport_lookup(
  country = NULL,
  city = NULL,
  distance = 0.1,
  ignore.case = FALSE
)
}
\arguments{
\item{country}{Name of the country.}

\item{city}{Name of the city.}

\item{distance}{Maximum distance allowed for a match between the country/city given, and that of the value in the data set.}

\item{ignore.case}{If \code{FALSE}, the check for city/country is case-sensitive. If \code{TRUE}, case is ignored.}
}
\value{
Data frame containing the country, city, country code, port code, latitude, and longitude.
}
\description{
Given a country and/or city, find the name and/or code of a seaport for use in the \code{ferry_emissions} function.
}
\examples{
# Look up the city of Aberdeen to find the port_code for it
seaport_lookup(city = "Aberdeen")
# Search for a country and city and it finds matches
seaport_lookup(country = "United", city = "borunemouth", ignore.case = TRUE)
}
