% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seaports.R
\docType{data}
\name{seaports}
\alias{seaports}
\title{Table of Seaport Data}
\format{
A data frame with 8736 rows and 7 variables:
\describe{
\item{latitude}{Latitude of the port}
\item{longitude}{Longitude of the port}
\item{country_name}{Name of the country with the port}
\item{country_code}{Code of the country name}
\item{port_code}{Code of the city port}
\item{city}{Name of the city with the port}
\item{id}{Row number}
}
}
\usage{
seaports
}
\description{
A dataset containing the country, country code, city, city code, and coordinates of seaports
}
\keyword{datasets}
