% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondary_emissions.R
\name{secondary_emissions}
\alias{secondary_emissions}
\title{Calculate CO2e emissions from secondary sources}
\usage{
secondary_emissions(item, cost, time = "per year", diet = "medium meat")
}
\arguments{
\item{item}{Item bought. Valid options are: \code{"IT equipment"}, \code{"restaurants"}, \code{"TV"}, \code{"radio"}, \code{"phone"}, \code{"paper based products"}, \code{"motor"}, \code{"manufactured goods"}, \code{"telephone/mobile call costs"}, \code{"banking/finance"}, \code{"insurance"}, \code{"education"}, \code{"recreational activities"}, \code{"textiles"}, \code{"food"}.}

\item{cost}{Amount spent on that item (in pounds)}

\item{time}{Time frame where the purchases took place. Options are \code{"per week"}, \code{"per month"}, \code{"per year"}.}

\item{diet}{Only if \code{item = "food"}. General diet of an individual. Options are \code{"high meat"}, \code{"medium meat"}, \code{"low meat"}, \code{"pescatarian"}, \code{"vegetarian"}, and \code{"vegan"}.
\code{"high meat"} suggests more than 100g of meat is consumed a day; \code{"medium meat"} for 50-100g/day; \code{"low meat"} for <50g/day.}
}
\value{
Tonnes of CO2 emissions from secondary sources in a year.
}
\description{
CO2e emissions from the manufacture, delivery, and disposal of products and services in a year. Emission values are calculated from carbonfootprint.com, which base the emissions values off DEFRA 2017 Supply Chain Factors.
}
\examples{
# Emissions for a new £1000 computer. 
secondary_emissions(item = "IT equipment", cost = 1000)
# Emissions for a high meat eater who spends £200 on food each month. 
secondary_emissions(item = "food", cost = 200, time = "per month")
}
