% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehicle_emissions.R
\name{vehicle_emissions}
\alias{vehicle_emissions}
\title{Calculate Emissions from Land-Travel}
\usage{
vehicle_emissions(
  distance,
  units = "miles",
  vehicle = "car",
  fuel = "petrol",
  size = "average",
  type = "average",
  taxi_type = "regular"
)
}
\arguments{
\item{distance}{Distance in Km or miles of the journey made (this can be calculated with other tools, such as google maps.).}

\item{units}{Whether the distance is in KM or miles.}

\item{vehicle}{Vehicle used for the journey. Options are "car", "motorbike", "taxi", "van", "bus", "coach", "tram", "tube". Note: bus, coach, tram, tube, are all per passenger}

\item{fuel}{}

\item{size}{}

\item{type}{}

\item{taxi_type}{}
}
\value{
T of CO2e burnt per mile
}
\description{
A function that calculates CO2 emissions on a journey on land.

Values to calculate emissions is UK DEFRA - see https://assets.publishing.service.gov.uk/government/uploads/system/uploads/attachment_data/file/990675/2021-ghg-conversion-factors-methodology.pdf

https://www.gov.uk/government/publications/greenhouse-gas-reporting-conversion-factors-2021
}
