% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vehicle_emissions.R
\name{vehicle_emissions}
\alias{vehicle_emissions}
\title{Calculate CO2e emissions from land-travel journeys}
\usage{
vehicle_emissions(
  distance,
  units = c("miles", "km"),
  num = 1,
  vehicle = c("car", "motorbike", "taxi", "van", "bus", "coach", "tram", "tube"),
  fuel = c("petrol", "diesel", "hybrid", "unknown", "hybrid electric",
    "battery electric"),
  size = c("average", "small", "medium", "large"),
  bus_type = c("local not London", "local London", "average"),
  taxi_type = c("regular", "black cab")
)
}
\arguments{
\item{distance}{Distance in km or miles of the journey made (this can be calculated with other tools, such as google maps.).}

\item{units}{Units for the distance travelled. Options are \code{"km"} or \code{"miles"}.}

\item{num}{Number of passengers if \code{vehicle} is one of \code{coach}, \code{tram}, or \code{tube}. Otherwise, number of vehicles used.}

\item{vehicle}{Vehicle used for the journey. Options are \code{"car"}, \code{"motorbike"}, \code{"taxi"}, \code{"van"}, \code{"bus"}, \code{"coach"}, \code{"tram"}, \code{"tube"}. Note: bus, coach, tram, tube, are all per passenger}

\item{fuel}{Fuel type used for the journey. For car, \code{"petrol"}, \code{"diesel"}, \code{"hybrid"}, \code{"unknown"}, \code{"hybrid electric"}, \code{"battery electric"} are options. For van, \code{"petrol"}, \code{"diesel"}, and \code{"battery electric"} are options.
\code{"hybrid electric"} and \code{"battery electric"} account for electricity kWh emissions.}

\item{size}{Size of vehicle for car, motorbike, and van.
Options are \code{"small"}, \code{"medium"}, \code{"large"}, or \code{"average"}.
For car: small denotes up to a 1.4L engine, unless diesel which is up to 1.7L engine. Medium denotes 1.4-2.0L for petrol cars, 1.7-2.0L for diesel cars. Large denotes 2.0L+ engine.
For motorbike, sizes denote upto 125cc, 125cc-500cc, 500cc+ respectively.}

\item{bus_type}{Options are \code{"local_nL"}, \code{"local_L"}, \code{"local"}, or \code{"average"}. These denote whether the bus is local but outside of London, local in London, local, or average.}

\item{taxi_type}{Whether a taxi is regular or black cab. Options are \code{"regular"}, \code{"black cab"}.}
}
\value{
Tonnes of CO2e emissions per mile travelled.
}
\description{
A function that calculates CO2e emissions on a journey on land.
}
\examples{
# Emissions for a 100 mile car journey
 vehicle_emissions(distance = 100)
# Emissions for a 100 mile motorbike journey where the motorbike is 500+cc
 vehicle_emissions(distance = 100, vehicle = "motorbike", size = "large")
}
